refactor(recordings): satisfy package policy gates
