+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/worker_pool.go
@@ -867,119 +867,3 @@ func cloneSessionCapabilities(value *workers.Capabilities) *workers.Capabilities
 	clone := *value
 	return &clone
 }
-
-const detachedAgentRunTranscriptSummaryLimit = 160
-
-func recordDetachedAgentRunResponse(
-	cfg *runtimeConfig,
-	request workers.ExecuteRequest,
-	result workers.ExecuteResult,
-	executeErr error,
-) {
-	if cfg == nil || cfg.eventHistory == nil || !runtimeRequestUsesAgentRun(cfg, request) {
-		return
-	}
-	recorder, ok := cfg.eventHistory.(recordings.WorkerEventRecorder)
-	if !ok || recorder == nil {
-		return
-	}
-	dispatchID := strings.TrimSpace(request.Correlation.DispatchID)
-	if dispatchID == "" {
-		return
-	}
-
-	transcript := make([]workers.AgentRunTranscriptEntry, 0, 3)
-	appendTranscriptEntry(&transcript, "system", request.Target.Prompt.SystemPrompt)
-	appendTranscriptEntry(&transcript, "user", request.Target.Prompt.UserMessage)
-	appendTranscriptEntry(&transcript, "assistant", primaryOutputText(result.Output.Primary))
-	safeDiagnostics := workers.SafeWorkDiagnosticsFromWorkDiagnostics(result.Diagnostics.ToWorkDiagnostics())
-	if safeDiagnostics == nil {
-		safeDiagnostics = &workers.SafeWorkDiagnostics{}
-	}
-	safeDiagnostics.AgentRun = &workers.SafeAgentRunDiagnostic{
-		ExecutionBehavior: workers.AgentRunExecutionBehavior,
-		Transcript:        transcript,
-	}
-	diagnostics, err := workers.SafeWorkDiagnosticsEventPayload(safeDiagnostics)
-	if err != nil {
-		return
-	}
-	outcome := result.Outcome
-	if outcome == "" && executeErr != nil {
-		outcome = workers.ExecutionOutcomeFailed
-	}
-	eventTime := cfg.clock.Now()
-	recorder.RecordAgentRunEvent(workers.AgentRunResponseEvent{
-		ID:                         fmt.Sprintf("factory-event/agent-run-response/%s", dispatchID),
-		DispatchID:                 dispatchID,
-		EventTime:                  eventTime,
-		Tick:                       detachedExecutionTick(request.Input.Dispatch.Execution),
-		DeclaredSecretJSONPointers: agentRunSecretJSONPointers(request, transcript),
-		Payload: workers.AgentRunResponseEventPayload{
-			AgentRunID:     fmt.Sprintf("%s/agent-run/1", dispatchID),
-			Diagnostics:    diagnostics,
-			DurationMillis: result.Metrics.Duration.Milliseconds(),
-			Outcome:        string(outcome),
-		},
-	})
-}
-
-func agentRunSecretJSONPointers(
-	request workers.ExecuteRequest,
-	transcript []workers.AgentRunTranscriptEntry,
-) []string {
-	if !hasDeclaredSecretInvocationParameter(request.Input.Invocation) {
-		return nil
-	}
-	pointers := make([]string, 0, len(transcript))
-	for index, entry := range transcript {
-		if entry.Role != "system" && entry.Role != "user" {
-			continue
-		}
-		pointers = append(pointers, fmt.Sprintf("/diagnostics/agentRun/transcript/%d/summary", index))
-	}
-	return pointers
-}
-
-func hasDeclaredSecretInvocationParameter(arguments work.InvocationArguments) bool {
-	for _, argument := range arguments.Arguments {
-		if argument.Sensitive {
-			return true
-		}
-	}
-	return false
-}
-
-func detachedExecutionTick(metadata work.ExecutionMetadata) int {
-	if metadata.CurrentTick != 0 {
-		return metadata.CurrentTick
-	}
-	return metadata.DispatchCreatedTick
-}
-
-func runtimeRequestUsesAgentRun(cfg *runtimeConfig, request workers.ExecuteRequest) bool {
-	lookup, ok := runtimeDefinitionLookup(cfg)
-	if !ok || lookup == nil {
-		return false
-	}
-	workstation, found := lookup.Workstation(strings.TrimSpace(request.Target.WorkstationName))
-	return found && workstation != nil && interfaces.IsAgentRunWorkstationType(workstation.Type)
-}
-
-func appendTranscriptEntry(
-	transcript *[]workers.AgentRunTranscriptEntry,
-	role string,
-	content string,
-) {
-	content = strings.TrimSpace(content)
-	if content == "" {
-		return
-	}
-	if len(content) > detachedAgentRunTranscriptSummaryLimit {
-		content = content[:detachedAgentRunTranscriptSummaryLimit] + "..."
-	}
-	*transcript = append(*transcript, workers.AgentRunTranscriptEntry{
-		Role:    role,
-		Summary: content,
-	})
-}
